requirements.txt
asgiref==3.7.2
Django==4.0.1
djangorestframework==3.14.0
psycopg2-binary==2.9.6
pytz==2023.3
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2023.3
dj-database-url==2.0.0
python-dotenv==0.21.0
django-filter==23.2
djangorestframework-filters
Pillow==10.0.0

